+++ backend-server/rabbitmqServer.php
@@ -3,6 +3,35 @@
 require_once('rabbitMQLib.inc');
 require_once('db.php.inc');
 
+function deleteUser($data) {
+	$response = array();
+	$response["status"] = false;
+	$response["messages"] = array();
+	
+	/* For Testing */
+	$status = true;	
+	echo "Performing deleteUser for ".$data['username'].PHP_EOL;
+	$response["status"] = $satus;
+	array_push($response['messages'], "Deleted user ".$data['username']);
+	
+	
+	/* to actually delete user from table */
+	/*$db_data = new Data();
+	$deleteUser_response = $db_data->deleteUser($data);
+	
+	# update response messages
+	foreach($deleteUser_response['messages'] as $msg){
+		array_push($response['messages'], $msg);
+	}
+	# check response status
+	if ($deleteUser_response['status'] === true) {
+		$response['status'] = true;
+	}*/
+	
+	print_r($response);
+	return $response;
+}
+
 function updateUser($data) {
 	$response = array();
 	$response["status"] = false;
@@ -40,7 +69,7 @@ function get_all_users_data() {
 		$response['status'] = true;
 		$response['data'] = $all_users_data_response['data'];
 	}
-	echo(count($response['data'])." Rows of data".PHP_EOL );
+	#echo(count($response['data'])." Rows of data".PHP_EOL );
 	return $response;
 }
 
@@ -111,6 +140,8 @@ function requestProcessor($request){
 				return get_all_users_data();
 			case "update-user":
 				return updateUser($request['data']);
+			case "delete-user":
+				return deleteUser($request['data']);
 		}
 	
 		return array("returnCode"=>"0", "message"=>"Server received request and processed");