+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v0.0.2
+* for admin login, display all user's from db Users table in scrollable html table
+	- can update username, email, role for users.
+	- username validation also occurs here
+	- if no changes were made but save-change button was clicked, the alert message says no changes made
+	- button to delete user, db queries the delete on table based on user-id & user-email 
+* registration form for new user
+	- username validation: "string.string" format to pass validation
+	- password and password-confirm validation
+	
 # v0.0.1
 * Beginning of CHANGELOG.md